backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
python-multipart==0.0.9
anthropic==0.25.0
slither-analyzer==0.10.0
solc-select==1.0.4
crytic-compile==0.3.6
openai==2.29.0
requests==2.31.0
